Rework docker build a bit:
To make it make sense, we had to let the user log in as their local
user.

We also didn't need all that git clone stuff, they can do that
themselves with a command, we now mount the main repo to /usr/share, and
we give them a convenience option to reclone that to their home
directory.